Android: 32-bit unicode support

+++ core/ui/gui.h
@@ -34,7 +34,7 @@ void gui_term();
 void gui_cancel_load();
 void gui_refresh_files();
 void gui_cheats();
-void gui_keyboard_input(u16 wc);
+void gui_keyboard_input(u32 wc);
 void gui_keyboard_inputUTF8(const std::string& s);
 void gui_keyboard_key(u8 keyCode, bool pressed);
 bool gui_keyboard_captured();
